backend/requirements_unified.txt
# AutoLabel Pro - Unified Server Requirements
# Install: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Image Processing
pillow>=10.2.0
numpy>=1.24.0,<2.0.0

# Deep Learning - PyTorch
# For CPU:
# torch>=2.1.0
# torchvision>=0.16.0

# For GPU (CUDA 11.8), install separately:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

torch>=2.1.0
torchvision>=0.16.0

# YOLO (YOLOv8 & YOLOv11)
ultralytics>=8.1.0

# Grounding DINO (from Hugging Face)
transformers>=4.37.0
huggingface-hub>=0.20.0
